app/analytics-backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
pytest==7.4.4
